accessability func changes
rewrote functions to make them clearer
